+++ R/Plots.R
@@ -466,3 +466,124 @@ plotSeasonality <- function(sccsModel,
     ggplot2::ggsave(fileName, plot, width = 7, height = 5, dpi = 400)
   return(plot)
 }
+
+#' Plot the calendar time ranges spanned by each observation period.
+#'
+#' @template StudyPopulation
+#' @param maxPersons  The maximum number of persons to plot. If there are more than this number of persons
+#'                    a random sample will be taken to avoid visual clutter.
+#' @param fileName    Name of the file where the plot should be saved, for example 'plot.png'.
+#'                    See the function [ggplot2::ggsave()] for supported file formats.
+#' @param title       Optional: the main title for the plot
+#'
+#' @details
+#' Plots a line per patient from their observation start to their observation end.
+#'
+#' @return
+#' A ggplot object. Use the [ggplot2::ggsave()] function to save to file in a different
+#' format.
+#'
+#' @export
+plotCalendarTimeSpans <- function(studyPopulation,
+                                  maxPersons = 10000,
+                                  title = NULL,
+                                  fileName = NULL) {
+  cases <- studyPopulation$cases %>%
+    mutate(endDate = .data$startDate + .data$endDay) %>%
+    select(.data$startDate, .data$endDate) %>%
+    arrange(.data$startDate, .data$endDate) %>%
+    mutate(rank = row_number())
+  if (nrow(cases) > maxPersons) {
+    warning("There are ", nrow(cases), " cases. Random sampling ", maxPersons, " cases.")
+    cases <- cases[sample.int(nrow(cases), maxPersons), ]
+  }
+  theme <- ggplot2::element_text(colour = "#000000", size = 12)
+  themeRA <- ggplot2::element_text(colour = "#000000", size = 12, hjust = 1)
+  plot <- ggplot2::ggplot(cases, ggplot2::aes(x = .data$startDate, xmin = .data$startDate, xmax = .data$endDate, y = rank)) +
+    ggplot2::geom_errorbarh(color = rgb(0, 0, 0.8)) +
+    ggplot2::scale_x_date("Calendar time") +
+    ggplot2::scale_y_continuous("Case rank") +
+    ggplot2::theme(panel.grid.minor = ggplot2::element_blank(),
+                   panel.background = ggplot2::element_rect(fill = "#FAFAFA", colour = NA),
+                   panel.grid.major.x = ggplot2::element_line(colour = "#AAAAAA", size = 0.2),
+                   panel.grid.major.y = ggplot2::element_blank(),
+                   axis.ticks = ggplot2::element_blank(),
+                   axis.text.y = themeRA,
+                   axis.text.x = theme,
+                   strip.text.x = theme,
+                   strip.background = ggplot2::element_blank(),
+                   plot.title = ggplot2::element_text(hjust = 0.5),
+                   legend.title = ggplot2::element_blank(),
+                   legend.position = "top")
+  if (!is.null(title)) {
+    plot <- plot + ggplot2::ggtitle(title)
+  }
+  # fileName <- "S:/temp/plot.png"
+  if (!is.null(fileName))
+    ggplot2::ggsave(fileName, plot, width = 7, height = 5, dpi = 400)
+  return(plot)
+}
+
+#' Plot the calendar time effect
+#'
+#' @details
+#' Plot the spline curve of the calendar time effect.
+#'
+#' @param sccsModel   An object of type \code{sccsModel} as created using the
+#'                    \code{\link{fitSccsModel}} function.
+#' @param rrLim       The limits on the incidence rate ratio scale in the plot.
+#' @param fileName    Name of the file where the plot should be saved, for example 'plot.png'. See the
+#'                    function \code{ggsave} in the ggplot2 package for supported file formats.
+#' @param title       Optional: the main title for the plot
+#'
+#' @return
+#' A Ggplot object. Use the ggsave function to save to file.
+#'
+#' @export
+plotCalendarTime <- function(sccsModel,
+                             rrLim = c(0.1, 10),
+                             title = NULL,
+                             fileName = NULL) {
+  estimates <- sccsModel$estimates
+  estimates <- estimates[estimates$covariateId >= 300 & estimates$covariateId < 400, ]
+  splineCoefs <- c(0, estimates$logRr)
+  calendarTimeKnots <- sccsModel$metaData$calendarTime$calendarTimeKnots
+  # calendarTimeKnots <-
+  calendarTime <- seq(min(calendarTimeKnots), max(calendarTimeKnots), length.out = 100)
+  seasonDesignMatrix <- cyclicSplineDesign(season, seasonKnots)
+  logRr <- apply(seasonDesignMatrix %*% splineCoefs, 1, sum)
+  logRr <- logRr - mean(logRr)
+  rr <- exp(logRr)
+  data <- data.frame(season = season, rr = rr)
+
+  breaks <- c(0.1, 0.25, 0.5, 1, 2, 4, 6, 8, 10)
+  seasonBreaks <- 1:12
+  theme <- ggplot2::element_text(colour = "#000000", size = 12)
+  themeRA <- ggplot2::element_text(colour = "#000000", size = 12, hjust = 1)
+  plot <- ggplot2::ggplot(data, ggplot2::aes(x = season, y = rr)) +
+    ggplot2::geom_hline(yintercept = breaks, colour = "#AAAAAA", lty = 1, size = 0.2) +
+    ggplot2::geom_line(color = rgb(0, 0, 0.8), alpha = 0.8, lwd = 1) +
+    ggplot2::scale_x_continuous("Month", breaks = seasonBreaks, labels = seasonBreaks) +
+    ggplot2::scale_y_continuous("Relative risk",
+                                limits = rrLim,
+                                trans = "log10",
+                                breaks = breaks,
+                                labels = breaks) +
+    ggplot2::theme(panel.grid.minor = ggplot2::element_blank(),
+                   panel.background = ggplot2::element_rect(fill = "#FAFAFA", colour = NA),
+                   panel.grid.major = ggplot2::element_blank(),
+                   axis.ticks = ggplot2::element_blank(),
+                   axis.text.y = themeRA,
+                   axis.text.x = theme,
+                   strip.text.x = theme,
+                   strip.background = ggplot2::element_blank(),
+                   plot.title = ggplot2::element_text(hjust = 0.5),
+                   legend.title = ggplot2::element_blank(),
+                   legend.position = "top")
+  if (!is.null(title)) {
+    plot <- plot + ggplot2::ggtitle(title)
+  }
+  if (!is.null(fileName))
+    ggplot2::ggsave(fileName, plot, width = 7, height = 5, dpi = 400)
+  return(plot)
+}