+++ src/RcppExports.cpp
@@ -5,9 +5,14 @@
 
 using namespace Rcpp;
 
+#ifdef RCPP_USE_GLOBAL_ROSTREAM
+Rcpp::Rostream<true>&  Rcpp::Rcout = Rcpp::Rcpp_cout_get();
+Rcpp::Rostream<false>& Rcpp::Rcerr = Rcpp::Rcpp_cerr_get();
+#endif
+
 // convertToSccs
-SEXP convertToSccs(const DataFrame& cases, const DataFrame& outcomes, const List& eras, const bool includeAge, const int ageOffset, const NumericMatrix& ageDesignMatrix, const bool includeSeason, const NumericMatrix& seasonDesignMatrix, const NumericVector& ageSeasonsCases, const List& covariateSettingsList, const bool eventDependentObservation, const List& censorModel, const bool scri, const int64_t controlIntervalId);
-RcppExport SEXP _SelfControlledCaseSeries_convertToSccs(SEXP casesSEXP, SEXP outcomesSEXP, SEXP erasSEXP, SEXP includeAgeSEXP, SEXP ageOffsetSEXP, SEXP ageDesignMatrixSEXP, SEXP includeSeasonSEXP, SEXP seasonDesignMatrixSEXP, SEXP ageSeasonsCasesSEXP, SEXP covariateSettingsListSEXP, SEXP eventDependentObservationSEXP, SEXP censorModelSEXP, SEXP scriSEXP, SEXP controlIntervalIdSEXP) {
+SEXP convertToSccs(const DataFrame& cases, const DataFrame& outcomes, const List& eras, const bool includeAge, const int ageOffset, const NumericMatrix& ageDesignMatrix, const bool includeSeason, const bool includeCalendarTime, const int calendarTimeOffset, const NumericMatrix& calendarTimeDesignMatrix, const NumericMatrix& seasonDesignMatrix, const NumericVector& timeCovariateCases, const List& covariateSettingsList, const bool eventDependentObservation, const List& censorModel, const bool scri, const int64_t controlIntervalId);
+RcppExport SEXP _SelfControlledCaseSeries_convertToSccs(SEXP casesSEXP, SEXP outcomesSEXP, SEXP erasSEXP, SEXP includeAgeSEXP, SEXP ageOffsetSEXP, SEXP ageDesignMatrixSEXP, SEXP includeSeasonSEXP, SEXP includeCalendarTimeSEXP, SEXP calendarTimeOffsetSEXP, SEXP calendarTimeDesignMatrixSEXP, SEXP seasonDesignMatrixSEXP, SEXP timeCovariateCasesSEXP, SEXP covariateSettingsListSEXP, SEXP eventDependentObservationSEXP, SEXP censorModelSEXP, SEXP scriSEXP, SEXP controlIntervalIdSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -18,14 +23,17 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< const int >::type ageOffset(ageOffsetSEXP);
     Rcpp::traits::input_parameter< const NumericMatrix& >::type ageDesignMatrix(ageDesignMatrixSEXP);
     Rcpp::traits::input_parameter< const bool >::type includeSeason(includeSeasonSEXP);
+    Rcpp::traits::input_parameter< const bool >::type includeCalendarTime(includeCalendarTimeSEXP);
+    Rcpp::traits::input_parameter< const int >::type calendarTimeOffset(calendarTimeOffsetSEXP);
+    Rcpp::traits::input_parameter< const NumericMatrix& >::type calendarTimeDesignMatrix(calendarTimeDesignMatrixSEXP);
     Rcpp::traits::input_parameter< const NumericMatrix& >::type seasonDesignMatrix(seasonDesignMatrixSEXP);
-    Rcpp::traits::input_parameter< const NumericVector& >::type ageSeasonsCases(ageSeasonsCasesSEXP);
+    Rcpp::traits::input_parameter< const NumericVector& >::type timeCovariateCases(timeCovariateCasesSEXP);
     Rcpp::traits::input_parameter< const List& >::type covariateSettingsList(covariateSettingsListSEXP);
     Rcpp::traits::input_parameter< const bool >::type eventDependentObservation(eventDependentObservationSEXP);
     Rcpp::traits::input_parameter< const List& >::type censorModel(censorModelSEXP);
     Rcpp::traits::input_parameter< const bool >::type scri(scriSEXP);
     Rcpp::traits::input_parameter< const int64_t >::type controlIntervalId(controlIntervalIdSEXP);
-    rcpp_result_gen = Rcpp::wrap(convertToSccs(cases, outcomes, eras, includeAge, ageOffset, ageDesignMatrix, includeSeason, seasonDesignMatrix, ageSeasonsCases, covariateSettingsList, eventDependentObservation, censorModel, scri, controlIntervalId));
+    rcpp_result_gen = Rcpp::wrap(convertToSccs(cases, outcomes, eras, includeAge, ageOffset, ageDesignMatrix, includeSeason, includeCalendarTime, calendarTimeOffset, calendarTimeDesignMatrix, seasonDesignMatrix, timeCovariateCases, covariateSettingsList, eventDependentObservation, censorModel, scri, controlIntervalId));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -114,7 +122,7 @@ END_RCPP
 }
 
 static const R_CallMethodDef CallEntries[] = {
-    {"_SelfControlledCaseSeries_convertToSccs", (DL_FUNC) &_SelfControlledCaseSeries_convertToSccs, 14},
+    {"_SelfControlledCaseSeries_convertToSccs", (DL_FUNC) &_SelfControlledCaseSeries_convertToSccs, 17},
     {"_SelfControlledCaseSeries_simulateSccsOutcomes", (DL_FUNC) &_SelfControlledCaseSeries_simulateSccsOutcomes, 9},
     {"_SelfControlledCaseSeries_testEwad", (DL_FUNC) &_SelfControlledCaseSeries_testEwad, 6},
     {"_SelfControlledCaseSeries_testEwid", (DL_FUNC) &_SelfControlledCaseSeries_testEwid, 6},